Rename health check classes to have some symmetry
