+++ src/main/java/com/palantir/docker/compose/connection/waiting/SingleServiceHealthCheck.java
@@ -19,6 +19,6 @@ package com.palantir.docker.compose.connection.waiting;
 import com.palantir.docker.compose.connection.Container;
 
 @FunctionalInterface
-public interface HealthCheck {
+public interface SingleServiceHealthCheck {
     SuccessOrFailure isServiceUp(Container container);
 }
